Poll Definition Export/Import
Implementation of export and import of a poll basic data (PollDefinition), including its text, responses and anonymity, in order to let moderator save his polls to load and use them later.

Created PollDefinition interface in poll model. It also includes the extraction of the poll creation elements to a separate component (PollCreationComponent)