Poll of different Types in Model
Added classes that represent the different poll types in the backend model. Also added custom exceptions and the PollResult class.

In order to preserve compatibility with the current version of the frontend, the previous version of the Poll class, the PollController and the polls' code in the OpenViduService remain unchanged.