Poll Creation in Backend
Created the PollBuilder class and one PollFactory class for each poll type in order to manage the creation of Poll objects in the backend.

In addition, the PollCreation class has been created in order to group all the information that the PollBuilder needs to create a poll, usually retrieved by the PollController.